move most common case to first
the most common case when parsing elements from the streaming API will
be tweets, thus, it should come first in the if/else branch rather than
last so that in more scenarios less checks will be performed.

this is a small difference but actually is highly visible when
streaming at high throughput.