twitter api: check for END OF TEXT
Twitter API seems to have updated to allow for better functionality
related to DMs but may have changed something else under the hood. We are
getting a return payload with END OF TEXT code:

`'\u0003\u0000\u0000\u0000\u0000\u0000\u0000\u0000\u0000\u0000'`

This PR checks for existence of this and continues with code instead of
throwing because of a `JSON.parse` failure.

Refs ttezel/twit#361